move login styles to modules
