+++ src/components/login-form/LoginForm.tsx
@@ -2,6 +2,7 @@ import {useForm} from "react-hook-form";
 import {joiResolver} from "@hookform/resolvers/joi";
 import {loginValidator} from "../../validators/login-validator.ts";
 import {useLogin} from "../../hooks/useLogin.ts";
+import styles from "./LoginForm.module.css";
 
 type LoginFormProps = {
     login: string;
@@ -13,19 +14,19 @@ export const LoginForm = () => {
     const {loginHandler, loginError} = useLogin();
 
     return (
-        <div className='p-10 border-2 border-gray-300 dark:border-gray-600 rounded-xl min-h-[400px] min-w-[400px]'>
-            <h1 className='text-3xl text-center mb-6'>Log In</h1>
-            {loginError && <p className='text-red-500 text-center mb-6'>{loginError}</p>}
-            <form onSubmit={handleSubmit(loginHandler)} className='flex flex-col items-center justify-center gap-6'>
-                <label className='flex flex-col items-start justify-center text-xs'>
-                    {errors.login ? <span className='text-red-500 pl-4'>{errors.login.message}</span> : <span className='pl-4'>Username: </span>}
-                    <input type="text" {...register('login')} className='h-[40px] min-w-[300px] text-base'/>
+        <div className={styles.formContainer}>
+            <h1 className={styles.formHeader}>Log In</h1>
+            {loginError && <p className={styles.loginError}>{loginError}</p>}
+            <form onSubmit={handleSubmit(loginHandler)} className={styles.form}>
+                <label className={styles.labelContainer}>
+                    {errors.login ? <span className={`${styles.label} ${styles.errorLabel}`}>{errors.login.message}</span> : <span className={styles.label}>Username: </span>}
+                    <input type="text" {...register('login')} className={styles.input}/>
                 </label>
-                <label className='flex flex-col items-start justify-center text-xs'>
-                    {errors.password ? <span className='text-red-500 pl-4'>{errors.password.message}</span> : <span className='pl-4'>Password: </span>}
-                    <input type="password" {...register('password')} className='h-[40px] min-w-[300px] text-base'/>
+                <label className={styles.labelContainer}>
+                    {errors.password ? <span className={`${styles.label} ${styles.errorLabel}`}>{errors.password.message}</span> : <span className={styles.label}>Password: </span>}
+                    <input type="password" {...register('password')} className={styles.input}/>
                 </label>
-                <button type="submit" disabled={!isValid} className=' text-gray-50 disabled:bg-gray-50 dark:disabled:bg-gray-900 disabled:border-2 disabled:border-gray-300 dark:disabled:border-gray-600 disabled:text-gray-300 dark:disabled:text-gray-600  rounded-2xl min-w-[300px] min-h-[40px] text-lg mt-4 enabled:cursor-pointer bg-indigo-600/90 enabled:hover:bg-indigo-600 enabled:hover:shadow-md enabled:hover:shadow-indigo-600/50'>LogIn</button>
+                <button type="submit" disabled={!isValid} className={styles.submitButton}>LogIn</button>
             </form>
         </div>
     );
