--- src/pages/LoginPage.tsx
@@ -1,9 +1,0 @@
-import {LoginForm} from "../components/login-form/LoginForm.tsx";
-
-export const LoginPage = () => {
-    return (
-        <div className='min-h-[60dvh] flex items-center justify-center'>
-            <LoginForm/>
-        </div>
-    );
-};
